services/web/requirements.txt
attrs==19.3.0
Click==7.0
Flask==1.1.1
Flask-Login==0.4.1
importlib-metadata==1.3.0
itsdangerous==1.1.0
Jinja2==2.10.3
kafka-python==1.4.7
MarkupSafe==1.1.1
more-itertools==8.0.2
packaging==20.0
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.1
pyparsing==2.4.6
pytest==5.3.2
six==1.13.0
wcwidth==0.1.8
Werkzeug==0.16.0
zipp==0.6.0
gunicorn==19.9.0
